requirements.txt
jupyter
matplotlib
pandas
plotly
python-dotenv
streamlit
yfinance